+++ lib/ddtrace/configuration/options.rb
@@ -1,4 +1,3 @@
-require 'ddtrace/configuration/option'
 require 'ddtrace/configuration/option_set'
 require 'ddtrace/configuration/option_definition'
 require 'ddtrace/configuration/option_definition_set'
@@ -24,22 +23,33 @@ module Datadog
         protected
 
         def option(name, meta = {}, &block)
-          options[name] = OptionDefinition.new(name, meta, &block).tap do
-            define_option_accessors(name)
+          builder = OptionDefinition::Builder.new(name, meta, &block)
+          options[name] = builder.to_definition.tap do
+            # Resolve and define helper functions
+            helpers = default_helpers(name).merge(builder.helpers)
+            define_helpers(helpers)
           end
         end
 
         private
 
-        def define_option_accessors(name)
-          option_name = name
+        def default_helpers(name)
+          option_name = name.to_sym
 
-          define_method(option_name) do
-            get_option(option_name)
-          end
+          {
+            option_name.to_sym => proc do
+              get_option(option_name)
+            end,
+            "#{option_name}=".to_sym => proc do |value|
+              set_option(option_name, value)
+            end
+          }
+        end
 
-          define_method("#{option_name}=") do |value|
-            set_option(option_name, value)
+        def define_helpers(helpers)
+          helpers.each do |name, block|
+            next unless block.is_a?(Proc)
+            define_method(name, &block)
           end
         end
       end
@@ -60,7 +70,11 @@ module Datadog
           options[name].get
         end
 
-        def to_h
+        def option_defined?(name)
+          self.class.options.key?(name)
+        end
+
+        def options_hash
           options.each_with_object({}) do |(key, _), hash|
             hash[key] = get_option(key)
           end
@@ -75,13 +89,13 @@ module Datadog
         def add_option(name)
           assert_valid_option!(name)
           definition = self.class.options[name]
-          Option.new(definition, self).tap do |option|
+          definition.build(self).tap do |option|
             options[name] = option
           end
         end
 
         def assert_valid_option!(name)
-          unless self.class.options.key?(name)
+          unless option_defined?(name)
             raise(InvalidOptionError, "#{self.class.name} doesn't define the option: #{name}")
           end
         end