+++ lib/ddtrace/contrib/action_view/instrumentation/template_renderer.rb
@@ -3,7 +3,6 @@ require 'ddtrace/contrib/action_view/ext'
 module Datadog
   module Contrib
     module ActionView
-      # Defines instrumentation for ActionView
       module Instrumentation
         # Instrumentation for template rendering
         module TemplateRenderer
@@ -83,7 +82,7 @@ module Datadog
             end
           end
 
-          # Rails >= 3.1 template rendering
+          # Shared code for Rails >= 3.1 template rendering
           module Rails31Plus
             def render(*args, &block)
               datadog_tracer.trace(
@@ -96,34 +95,36 @@ module Datadog
 
             def render_template(*args)
               begin
-                # arguments based on render_template signature (stable since Rails 3.2)
-                template = args[0]
-                layout_name = args[1]
+                template, layout_name = datadog_parse_args(*args)
 
-                # update the tracing context with computed values before the rendering
-                template_name = template.try('identifier')
-                template_name = Utils.normalize_template_name(template_name)
-                layout = layout_name.try(:[], 'virtual_path')
-
-                if template_name
-                  active_datadog_span.set_tag(
-                    Ext::TAG_TEMPLATE_NAME,
-                    template_name
-                  )
-                end
-
-                if layout
-                  active_datadog_span.set_tag(
-                    Ext::TAG_LAYOUT,
-                    layout
-                  )
-                end
+                datadog_render_template(template, layout_name)
               rescue StandardError => e
                 Datadog::Tracer.log.debug(e.message)
               end
 
               # execute the original function anyway
               super(*args)
+            end
+
+            def datadog_render_template(template, layout_name)
+              # update the tracing context with computed values before the rendering
+              template_name = template.try('identifier')
+              template_name = Utils.normalize_template_name(template_name)
+              layout = layout_name.try(:[], 'virtual_path') # Proc can be called without parameters since Rails 6
+
+              if template_name
+                active_datadog_span.set_tag(
+                  Ext::TAG_TEMPLATE_NAME,
+                  template_name
+                )
+              end
+
+              if layout
+                active_datadog_span.set_tag(
+                  Ext::TAG_LAYOUT,
+                  layout
+                )
+              end
             end
 
             private
@@ -141,49 +142,23 @@ module Datadog
               self.active_datadog_span = nil
             end
           end
-        end
 
-        # Instrumentation for partial rendering
-        module PartialRenderer
-          def render(*args, &block)
-            datadog_tracer.trace(
-              Ext::SPAN_RENDER_PARTIAL,
-              span_type: Datadog::Ext::HTTP::TEMPLATE
-            ) do |span|
-              with_datadog_span(span) { super(*args) }
+          # Rails >= 3.1 && < 6 template rendering
+          module Rails31To5
+            include Rails31Plus
+
+            def datadog_parse_args(template, layout_name, *args)
+              [template, layout_name]
             end
           end
 
-          def render_partial(*args)
-            begin
-              template_name = Utils.normalize_template_name(@template.try('identifier'))
-              if template_name
-                active_datadog_span.set_tag(
-                  Ext::TAG_TEMPLATE_NAME,
-                  template_name
-                )
-              end
-            rescue StandardError => e
-              Datadog::Tracer.log.debug(e.message)
+          # Rails >= 6 template rendering
+          module Rails6Plus
+            include Rails31Plus
+
+            def datadog_parse_args(view, template, layout_name, *args)
+              [template, layout_name]
             end
-
-            # execute the original function anyway
-            super(*args)
-          end
-
-          private
-
-          attr_accessor :active_datadog_span
-
-          def datadog_tracer
-            Datadog.configuration[:action_view][:tracer]
-          end
-
-          def with_datadog_span(span)
-            self.active_datadog_span = span
-            yield
-          ensure
-            self.active_datadog_span = nil
           end
         end
       end