+++ lib/ddtrace/contrib/active_model_serializers/event.rb
@@ -21,7 +21,7 @@ module Datadog
           end
 
           def tracer
-            configuration[:tracer]
+            -> { configuration[:tracer] }
           end
 
           def configuration