+++ lib/ddtrace/contrib/grape/instrumentation.rb
@@ -4,48 +4,8 @@ module Datadog
       # Instrumentation for Grape::Endpoint
       module Instrumentation
         def self.included(base)
-          if Gem::Version.new(RUBY_VERSION) < Gem::Version.new('2.0.0')
-            base.class_eval do
-              # Class methods
-              singleton_class.send(:include, ClassMethodsCompatibility)
-              singleton_class.send(:include, ClassMethods)
-
-              # Instance methods
-              include InstanceMethodsCompatibility
-              include InstanceMethods
-            end
-          else
-            base.singleton_class.send(:prepend, ClassMethods)
-            base.send(:prepend, InstanceMethods)
-          end
-        end
-
-        # Compatibility shim for Rubies not supporting `.prepend`
-        module ClassMethodsCompatibility
-          def self.included(base)
-            base.class_eval do
-              alias_method :generate_api_method_without_datadog, :generate_api_method
-              remove_method :generate_api_method
-            end
-          end
-
-          def generate_api_method(*args, &block)
-            generate_api_method_without_datadog(*args, &block)
-          end
-        end
-
-        # Compatibility shim for Rubies not supporting `.prepend`
-        module InstanceMethodsCompatibility
-          def self.included(base)
-            base.class_eval do
-              alias_method :run_without_datadog, :run
-              remove_method :run
-            end
-          end
-
-          def run(*args, &block)
-            run_without_datadog(*args, &block)
-          end
+          base.singleton_class.send(:prepend, ClassMethods)
+          base.send(:prepend, InstanceMethods)
         end
 
         # ClassMethods - implementing instrumentation