+++ lib/ddtrace/contrib/racecar/event.rb
@@ -26,7 +26,7 @@ module Datadog
           end
 
           def tracer
-            configuration[:tracer]
+            -> { configuration[:tracer] }
           end
 
           def configuration