+++ spec/ddtrace/contrib/active_record/multi_db_spec.rb
@@ -98,9 +98,8 @@ RSpec.describe 'ActiveRecord multi-database implementation' do
 
     context 'a Symbol that matches a configuration' do
       context 'when ActiveRecord has configurations' do
-        before(:each) do
-          # Stub ActiveRecord::Base, to pretend its been configured
-          allow(ActiveRecord::Base).to receive(:configurations).and_return(
+        let(:database_configuration) do
+          {
             'gadget' => {
               'encoding' => 'utf8',
               'adapter' => 'mysql2',
@@ -116,7 +115,21 @@ RSpec.describe 'ActiveRecord multi-database implementation' do
               'timeout' => 5000,
               'database' => ':memory:'
             }
-          )
+          }
+        end
+
+        let(:database_configuration_object) do
+          if defined?(ActiveRecord::DatabaseConfigurations)
+            ActiveRecord::DatabaseConfigurations.new(database_configuration)
+          else
+            # Before Rails 6, database configuration was a plain Hash
+            database_configuration
+          end
+        end
+
+        before(:each) do
+          # Stub ActiveRecord::Base, to pretend its been configured
+          allow(ActiveRecord::Base).to receive(:configurations).and_return(database_configuration_object)
 
           Datadog.configure do |c|
             c.use :active_record, describes: :gadget do |gadget_db|